Introduce Publications as injectable queries
On the client, Publications have .subscribe() which
subscribes to the data source on either the current template or the whole app.

On the server, Publications are published via a /dust/publication
publication.

In any environment the .find() method can be used to query
for the data the publication would contain.